hw9-bookshelf-spring-jdbc: refactored service with tests
