+++ src/main/java/com/otus/bookstore/dao/impl/AuthorDaoJdbc.java
@@ -1,8 +1,8 @@
 package com.otus.bookstore.dao.impl;
 
 import com.otus.bookstore.dao.AuthorDao;
-import com.otus.bookstore.dao.impl.query.author.*;
 import com.otus.bookstore.model.Author;
+import com.otus.bookstore.service.author.*;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
