+++ src/main/java/com/otus/bookstore/service/author/AuthorInsertQuery.java
@@ -1,4 +1,4 @@
-package com.otus.bookstore.dao.impl.query.author;
+package com.otus.bookstore.service.author;
 
 import org.springframework.jdbc.core.SqlParameter;
 import org.springframework.jdbc.object.SqlUpdate;
