+++ src/main/java/com/otus/bookstore/service/book/BookUpdateQuery.java
@@ -1,4 +1,4 @@
-package com.otus.bookstore.dao.impl.query.book;
+package com.otus.bookstore.service.book;
 
 import org.springframework.jdbc.core.SqlParameter;
 import org.springframework.jdbc.object.SqlUpdate;
@@ -6,11 +6,12 @@ import org.springframework.jdbc.object.SqlUpdate;
 import javax.sql.DataSource;
 import java.sql.Types;
 
-public class BookInsertQuery extends SqlUpdate {
-    static final String SQL_INSERT_BOOK = "INSERT INTO book (title, description, price, author_id, genre_id) VALUES (:title, :description, :price, :author_id, :genre_id)";
+public class BookUpdateQuery extends SqlUpdate {
+    private static final String SQL_UPDATE_BOOK = "UPDATE book SET title = :title, description = :description, price = :price, author_id = :author_id, genre_id = :genre_id WHERE id = :id";
 
-    public BookInsertQuery(DataSource dataSource) {
-        super(dataSource, SQL_INSERT_BOOK);
+    public BookUpdateQuery(DataSource dataSource) {
+        super(dataSource, SQL_UPDATE_BOOK);
+        super.declareParameter(new SqlParameter("id", Types.INTEGER));
         super.declareParameter(new SqlParameter("title", Types.VARCHAR));
         super.declareParameter(new SqlParameter("description", Types.VARCHAR));
         super.declareParameter(new SqlParameter("price", Types.NUMERIC));
@@ -18,5 +19,6 @@ public class BookInsertQuery extends SqlUpdate {
         super.declareParameter(new SqlParameter("genre_id", Types.INTEGER));
         super.setGeneratedKeysColumnNames("id");
         super.setReturnGeneratedKeys(true);
+        super.compile();
     }
 }
