+++ src/main/java/com/otus/bookstore/service/genre/GenreDeleteAllQuery.java
@@ -1,4 +1,4 @@
-package com.otus.bookstore.dao.impl.query.genre;
+package com.otus.bookstore.service.genre;
 
 import org.springframework.jdbc.object.SqlUpdate;
